// services/bff/src/modules/interview/enum/interview.enum.ts
export enum InterviewType {
    REAL = "REAL",
    PRACTICE = "PRACTICE",
}

export enum InterviewRole {
    TECH = "TECH",
    HR = "HR",
    LEADER = "LEADER",
}

export enum InterviewPersonality {
    PRESSURE = "PRESSURE",
    COMFORTABLE = "COMFORTABLE",
    RANDOM = "RANDOM",
}

export enum InterviewStatus {
    READY = "READY",
    IN_PROGRESS = "IN_PROGRESS",
    COMPLETED = "COMPLETED",
    CANCELLED = "CANCELLED",
    PAUSED = "PAUSED",
}

export enum InterviewStage {
    WAITING = "WAITING",
    GREETING = "GREETING",
    CANDIDATE_GREETING = "CANDIDATE_GREETING",
    INTERVIEWER_INTRO = "INTERVIEWER_INTRO",
    SELF_INTRO_PROMPT = "SELF_INTRO_PROMPT",
    SELF_INTRO = "SELF_INTRO",
    IN_PROGRESS = "IN_PROGRESS",
    LAST_QUESTION_PROMPT = "LAST_QUESTION_PROMPT",
    LAST_ANSWER = "LAST_ANSWER",
    CLOSING_GREETING = "CLOSING_GREETING",
    COMPLETED = "COMPLETED",
}